Fix runtime/int issues.
This CL fixes up various runtime/int types to be more specific.

Bug: dawn:1339
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/86865
Commit-Queue: Dan Sinclair <dsinclair@chromium.org>
Auto-Submit: Dan Sinclair <dsinclair@chromium.org>

+++ src/dawn/common/BitSetIterator.h
@@ -22,7 +22,8 @@
 #include "dawn/common/Math.h"
 #include "dawn/common/UnderlyingType.h"
 
-// This is ANGLE's BitSetIterator class with a customizable return type
+// This is ANGLE's BitSetIterator class with a customizable return type.
+// Types have been updated to be more specific.
 // TODO(crbug.com/dawn/306): it could be optimized, in particular when N <= 64
 
 template <typename T>
@@ -53,12 +54,12 @@ class BitSetIterator final {
         }
 
       private:
-        unsigned long getNextBit();
+        uint32_t getNextBit();
 
         static constexpr size_t kBitsPerWord = sizeof(uint32_t) * 8;
         std::bitset<N> mBits;
-        unsigned long mCurrentBit;
-        unsigned long mOffset;
+        uint32_t mCurrentBit;
+        uint32_t mOffset;
     };
 
     Iterator begin() const {
@@ -92,7 +93,7 @@ BitSetIterator<N, T>::Iterator::Iterator(const std::bitset<N>& bits)
     if (bits.any()) {
         mCurrentBit = getNextBit();
     } else {
-        mOffset = static_cast<unsigned long>(roundUp(N, kBitsPerWord));
+        mOffset = static_cast<uint32_t>(roundUp(N, kBitsPerWord));
     }
 }
 
@@ -115,7 +116,7 @@ bool BitSetIterator<N, T>::Iterator::operator!=(const Iterator& other) const {
 }
 
 template <size_t N, typename T>
-unsigned long BitSetIterator<N, T>::Iterator::getNextBit() {
+uint32_t BitSetIterator<N, T>::Iterator::getNextBit() {
     static std::bitset<N> wordMask(std::numeric_limits<uint32_t>::max());
 
     while (mOffset < N) {
