+++ src/dawn/common/HashUtils.h
@@ -75,7 +75,7 @@ void HashCombine(size_t* hash, const T& value, const Args&... args) {
 #if defined(_GLIBCXX_DEBUG)
 template <size_t N>
 size_t Hash(const std::bitset<N>& value) {
-    constexpr size_t kWindowSize = sizeof(unsigned long long);
+    constexpr size_t kWindowSize = sizeof(uint64_t);
 
     std::bitset<N> bits = value;
     size_t hash = 0;
